src/lib/ai/content-optimizer.ts
import { OpenAI } from 'openai';
import { enhancedCache } from '../enhanced-cache';
import { LinkBioProfile } from '@/types/link-bio';

const openai = new OpenAI({
  apiKey: process.env.OPENAI_API_KEY,
});

interface OptimizationSuggestion {
  type: 'bio' | 'link' | 'seo' | 'accessibility';
  originalContent: string;
  suggestedContent: string;
  reason: string;
  impact: 'high' | 'medium' | 'low';
  metrics?: Record<string, number>;
}

interface ContentScore {
  clarity: number;
  engagement: number;
  seo: number;
  accessibility: number;
  overall: number;
}

export class ContentOptimizer {
  private static instance: ContentOptimizer;

  private constructor() {}

  static getInstance(): ContentOptimizer {
    if (!ContentOptimizer.instance) {
      ContentOptimizer.instance = new ContentOptimizer();
    }
    return ContentOptimizer.instance;
  }

  // Optimize bio description
  async optimizeBio(profile: LinkBioProfile): Promise<OptimizationSuggestion[]> {
    const cacheKey = `bio-optimization:${profile.id}`;
    const cached = await enhancedCache.get(cacheKey);
    if (cached) return cached;

    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: `You are an expert content optimizer. Analyze and improve the bio for maximum engagement and clarity.
                   Consider SEO, readability, and emotional appeal.`
        },
        {
          role: "user",
          content: `Optimize this bio: "${profile.bio}"
                   Consider:
                   1. Target audience
                   2. Professional tone
                   3. Keywords
                   4. Call to action
                   5. Character limits`
        }
      ],
      temperature: 0.7,
    });

    const suggestions = this.parseSuggestions(completion.choices[0].message.content);
    await enhancedCache.set(cacheKey, suggestions, 3600); // Cache for 1 hour
    return suggestions;
  }

  // Generate optimized link titles
  async optimizeLinkTitles(profile: LinkBioProfile): Promise<OptimizationSuggestion[]> {
    const suggestions: OptimizationSuggestion[] = [];
    
    for (const link of profile.customLinks) {
      const completion = await openai.chat.completions.create({
        model: "gpt-4",
        messages: [
          {
            role: "system",
            content: "Optimize link titles for maximum click-through rate while maintaining clarity and trust."
          },
          {
            role: "user",
            content: `Optimize this link title: "${link.title}"
                     URL: ${link.url}
                     Consider:
                     1. Click-through potential
                     2. Clarity
                     3. Trust signals
                     4. Action words`
          }
        ],
        temperature: 0.7,
      });

      const suggestion = {
        type: 'link' as const,
        originalContent: link.title,
        suggestedContent: completion.choices[0].message.content,
        reason: "Improved click-through potential and clarity",
        impact: 'medium' as const,
      };

      suggestions.push(suggestion);
    }

    return suggestions;
  }

  // Generate SEO meta tags
  async generateSEOTags(profile: LinkBioProfile): Promise<{
    title: string;
    description: string;
    keywords: string[];
    openGraph: Record<string, string>;
  }> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Generate optimized SEO meta tags based on profile content."
        },
        {
          role: "user",
          content: `Generate SEO tags for:
                   Name: ${profile.name}
                   Bio: ${profile.bio}
                   Links: ${profile.customLinks.map(l => l.title).join(', ')}`
        }
      ],
      temperature: 0.7,
    });

    // Parse and structure SEO tags
    return {
      title: '',
      description: '',
      keywords: [],
      openGraph: {},
    };
  }

  // Score content quality
  async scoreContent(content: string): Promise<ContentScore> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Analyze content quality and provide numerical scores."
        },
        {
          role: "user",
          content: `Score this content: "${content}"
                   Provide scores (0-100) for:
                   1. Clarity
                   2. Engagement
                   3. SEO
                   4. Accessibility`
        }
      ],
      temperature: 0.3,
    });

    // Parse scores from response
    return {
      clarity: 0,
      engagement: 0,
      seo: 0,
      accessibility: 0,
      overall: 0,
    };
  }

  // Generate A/B test variations
  async generateABVariations(content: string, count: number = 3): Promise<string[]> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Generate A/B test variations maintaining the core message but varying style and approach."
        },
        {
          role: "user",
          content: `Generate ${count} variations of: "${content}"
                   Maintain core message but vary:
                   1. Tone
                   2. Structure
                   3. Call to action
                   4. Word choice`
        }
      ],
      temperature: 0.9,
    });

    return this.parseVariations(completion.choices[0].message.content);
  }

  // Generate accessibility improvements
  async generateAccessibilityImprovements(profile: LinkBioProfile): Promise<OptimizationSuggestion[]> {
    const completion = await openai.chat.completions.create({
      model: "gpt-4",
      messages: [
        {
          role: "system",
          content: "Analyze and suggest improvements for web accessibility compliance."
        },
        {
          role: "user",
          content: `Analyze for accessibility:
                   Bio: ${profile.bio}
                   Links: ${profile.customLinks.map(l => l.title).join(', ')}
                   Consider:
                   1. Screen reader compatibility
                   2. Color contrast
                   3. Clear navigation
                   4. Semantic structure`
        }
      ],
      temperature: 0.5,
    });

    return this.parseSuggestions(completion.choices[0].message.content);
  }

  private parseSuggestions(response: string): OptimizationSuggestion[] {
    // Implementation of suggestion parsing
    return [];
  }

  private parseVariations(response: string): string[] {
    // Implementation of variation parsing
    return [];
  }
}

export const contentOptimizer = ContentOptimizer.getInstance();
